+++ cluster/images/docs/src/release_notes.rst
@@ -15,6 +15,8 @@ Release Notes
     Note that not all fields in the Scan UI are hooked up to fetch data in the backend yet.
     Ones that should work at this point are the as-of round in the top-right corner, and the Validator and App leaderboards.
 
+* Removed the ``svc-client`` config parameter from the SV app. The SVC app is no longer used for SV onboarding and initialization.
+
 2023-06-11
 ----------